Refactor payment method filtering logic
Moved filterActivePaymentMethods from zuora module to handler, and updated getPaymentMethods to use a generic return type. This improves separation of concerns and allows for more flexible schema usage.
